--- modules/store/src/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from "./fs";
