+++ modules/store/src/index.ts
@@ -8,12 +8,5 @@ export {
   WrappedMemoryStorage,
   WrappedPostgresStorage,
 } from "./wrappers";
-export {
-  getDirectoryFiles,
-  isDirectory,
-} from "./helpers";
-export {
-  storeDefaults,
-  storeKeys,
-  storePaths,
-} from "./constants";
+export { getDirectoryFiles, isDirectory } from "./helpers";
+export { storeDefaults, storeKeys, storePaths } from "./constants";
