+++ modules/store/src/types.ts
@@ -1,0 +1,9 @@
+export interface WrappedStorage {
+  getItem<T = any>(key: string): Promise<T | undefined>;
+  setItem<T = any>(key: string, value: T): Promise<void>;
+  removeItem(key: string): Promise<void>;
+  getKeys(): Promise<string[]>;
+  getEntries(): Promise<[string, any][]>;
+  // generates a key for related subject strings
+  getKey(...args: string[]): string;
+}
