+++ modules/store/src/wrappers/fileStorage.ts
@@ -1,4 +1,3 @@
-import { WrappedStorage } from "@connext/types";
 import { safeJsonParse, safeJsonStringify } from "@connext/utils";
 import writeFileAtomic from "write-file-atomic";
 
@@ -11,6 +10,7 @@ import {
   sanitizeExt,
 } from "../helpers";
 import { storeDefaults } from "../constants";
+import { WrappedStorage } from "../types";
 
 export class FileStorage implements WrappedStorage {
   constructor(
